Optimizer plugin: rewrite for sourcmaps support
Hook the plugin more deeply into the compilation process and make changes
on the assets in a sourcemap compatible fashion. Additionally, as the plugin
does not rely on dist files anymore, it is now also compatible with watch mode
which allows for testing translation optimizations in the normal dev cycle.