optimizer plugin: augment translation file hashes with source language contents
Augment content hashes of translation files with contents of the source language file. This is
because the content of the compiled translation file does not only depend on the translation po file,
but also on the source language file for added fallback translations and for translation indices.
Otherwise, the file hash would not update on source language changes, which can lead to cached files
being used which don't have the correct translations or wrong/missing indices.