optimizer plugin: support for Webpack eval devtool settings
For example eval, eval-cheap-source-map, eval-cheap-module-source-map, eval-source-map etc.
An eval option is the default if no devtool setting is specifically set. The eval options
wrap the code into a string literal which is evaluated by an eval call. Within the string
literal, string special characters quote, backslash and newline are escaped which can now
be handled by the optimizer plugin.